# [[ Perf ]] Make container resolution more efficient