--- a/engine/src/handler.cpp
+++ b/engine/src/handler.cpp
@@ -356,9 +356,8 @@ Exec_stat MCHandler::exec(MCExecContext& ctxt, MCParameter *plist)
 				}
                 
                 MCVariable *t_new_var;
-                /* UNCHECKED */ newparams[i] = new(nothrow) MCContainer;
-				/* UNCHECKED */ MCVariable::createwithname(i < npnames ? pinfo[i] . name : kMCEmptyName, t_new_var);
-                /* UNCHECKED */ MCContainer::createwithvariable(t_new_var, *newparams[i]);
+                /* UNCHECKED */ MCVariable::createwithname(i < npnames ? pinfo[i] . name : kMCEmptyName, t_new_var);
+                /* UNCHECKED */ newparams[i] = new(nothrow) MCContainer(t_new_var);
                 
 				newparams[i]->give_value(ctxt, t_value);
 			}
@@ -378,9 +377,8 @@ Exec_stat MCHandler::exec(MCExecContext& ctxt, MCParameter *plist)
 				break;
 			}
             MCVariable *t_new_var;
-            /* UNCHECKED */ newparams[i] = new(nothrow) MCContainer;
             /* UNCHECKED */ MCVariable::createwithname(i < npnames ? pinfo[i] . name : kMCEmptyName, t_new_var);
-            /* UNCHECKED */ MCContainer::createwithvariable(t_new_var, *newparams[i]);
+            /* UNCHECKED */ newparams[i] = new(nothrow) MCContainer(t_new_var);
 		}
 	}
 	if (err)