--- a/engine/src/variable.cpp
+++ b/engine/src/variable.cpp
@@ -1038,26 +1038,26 @@ MCVarref *MCVariable::newvarref(void)
 
 ////////////////////////////////////////////////////////////////////////////////
 
-static MCAutoPointer<MCNameRef[]>
-__join_paths(MCSpan<MCNameRef> p_base,
-             MCSpan<MCNameRef> p_extra)
-{
-    MCAutoPointer<MCNameRef[]> t_result =
-            new (nothrow) MCNameRef[p_base.size() + p_extra.size()];
-    if (t_result)
-    {
-        int t_count = 0;
-        for (MCNameRef t_name : p_base)
-            t_result[t_count++] = t_name;
-        for (MCNameRef t_name : p_extra)
-            t_result[t_count++] = t_name;
-    }
-    
-    return t_result;
-}
-
 MCContainer::~MCContainer(void)
 {
+    switch(m_path_length)
+    {
+    case 4:
+        MCValueRelease(m_short_path[3]);
+    case 3:
+        MCValueRelease(m_short_path[2]);
+    case 2:
+        MCValueRelease(m_short_path[1]);
+    case 1:
+        MCValueRelease(m_short_path[0]);
+    case 0:
+        break;
+    default:
+        for(unsigned i = 0; i < m_path_length; i++)
+            MCValueRelease(m_long_path[i]);
+        MCMemoryDeleteArray(m_long_path);
+        break;
+    }
 }
 
 bool MCContainer::eval(MCExecContext& ctxt, MCValueRef& r_value)
@@ -1065,32 +1065,9 @@ bool MCContainer::eval(MCExecContext& ctxt, MCValueRef& r_value)
     return m_variable -> eval(ctxt, getpath(), r_value);
 }
 
-bool MCContainer::eval_on_path(MCExecContext& ctxt, MCSpan<MCNameRef> p_path, MCValueRef& r_value)
-{
-    MCAutoPointer<MCNameRef[]> t_full_path = __join_paths(m_path.Span(), p_path);
-    
-    if (!t_full_path)
-        return false;
-
-    return m_variable->eval(ctxt,
-                            {*t_full_path, p_path.size() + m_path.Span().size()},
-                            r_value);
-}
-
 bool MCContainer::set(MCExecContext& ctxt, MCValueRef p_value, MCVariableSettingStyle p_setting)
 {
 	return m_variable -> set(ctxt, p_value, getpath(), p_setting);
-}
-
-bool MCContainer::set_on_path(MCExecContext& ctxt, MCSpan<MCNameRef> p_path, MCValueRef p_value)
-{
-    MCAutoPointer<MCNameRef[]> t_full_path = __join_paths(m_path.Span(), p_path);
-    
-    if (!t_full_path)
-        return false;
-    
-    return m_variable->set(ctxt, p_value,
-                           {*t_full_path, p_path.size() + m_path.Span().size()});
 }
 
 bool MCContainer::eval_ctxt(MCExecContext& ctxt, MCExecValue& r_value)
@@ -1120,12 +1097,12 @@ bool MCContainer::remove(MCExecContext& ctxt)
 
 bool MCContainer::set_valueref(MCValueRef p_value)
 {
-	return m_variable -> setvalueref(getpath(), m_case_sensitive, p_value);
+	return m_variable -> setvalueref(getpath(), false, p_value);
 }
 
 MCValueRef MCContainer::get_valueref()
 {
-	return m_variable -> getvalueref(getpath(), m_case_sensitive);
+	return m_variable -> getvalueref(getpath(), false);
 }
 
 bool MCContainer::clear(void)
@@ -1135,7 +1112,7 @@ bool MCContainer::clear(void)
 
 bool MCContainer::set_real(double p_real)
 {
-    if (m_path.Count() == 0)
+    if (m_path_length == 0)
     {
         m_variable -> setnvalue(p_real);
         return true;
@@ -1149,28 +1126,12 @@ bool MCContainer::set_real(double p_real)
 
 MCSpan<MCNameRef> MCContainer::getpath()
 {
-    return m_path.Span();
-}
-
-bool MCContainer::createwithvariable(MCVariable *p_var, MCContainer& r_container)
-{
-	r_container.m_variable = p_var;
-    r_container.m_path.Reset();
-	r_container.m_case_sensitive = false;
-	return true;
-}
-
-bool MCContainer::createwithpath(MCVariable *p_var, MCNameRef *p_path, uindex_t p_length, MCContainer& r_container)
-{
-	r_container.m_variable = p_var;
-	r_container.m_path.Give(p_path, p_length);
-	r_container.m_case_sensitive = false;
-	return true;
-}
-
-bool MCContainer::copywithpath(MCContainer *p_container, MCNameRef *p_path, uindex_t p_length, MCContainer& r_container)
-{
-	return createwithpath(p_container -> m_variable, p_path, p_length, r_container);
+    MCNameRef *t_path = nullptr;
+    if (m_path_length <= sizeof(m_short_path) / sizeof(m_short_path[0]))
+        t_path = m_short_path;
+    else
+        t_path = m_long_path;
+    return MCMakeSpan(t_path, m_path_length);
 }
 
 ////////////////////////////////////////////////////////////////////////////////
@@ -1367,11 +1328,100 @@ bool MCVarref::dofree(MCExecContext& ctxt)
 	return t_container.remove(ctxt);
 }
 
+class MCContainerBuilder
+{
+public:
+    MCContainerBuilder(MCContainer& p_target)
+        : m_container(p_target)
+    {
+    }
+    
+    bool AppendValueRef(MCExecContext& ctxt, MCValueRef p_value)
+    {
+        switch(MCValueGetTypeCode(p_value))
+        {
+            case kMCValueTypeCodeName:
+                return AppendNameRef((MCNameRef)p_value);
+            case kMCValueTypeCodeArray:
+                if (MCArrayIsSequence((MCArrayRef)p_value))
+                    return AppendValueSeq(ctxt, (MCArrayRef)p_value);
+                ctxt . LegacyThrow(EE_VARIABLE_BADINDEX);
+                return false;
+            default:
+                break;
+        }
+                
+        MCNewAutoNameRef t_name;
+        if (!ctxt.ConvertToName(p_value, &t_name))
+        {
+            ctxt.LegacyThrow(EE_VARIABLE_BADINDEX);
+            return false;
+        }
+        
+        return AppendNameRef(*t_name);
+    }
+    
+    bool AppendValueSeq(MCExecContext& ctxt, MCArrayRef p_seq)
+    {
+        uindex_t t_length = MCArrayGetCount(p_seq);
+        for(uindex_t t_index = 1; t_index <= t_length; t_index += 1)
+        {
+            MCValueRef t_value_fetched;
+            if (!MCArrayFetchValueAtIndex(p_seq, t_index, t_value_fetched))
+                return false;
+            
+            if (!AppendValueRef(ctxt, t_value_fetched))
+                return false;
+        }
+        return true;
+    }
+    
+    bool AppendNameSpan(MCSpan<MCNameRef> p_path)
+    {
+        for (MCNameRef t_name : p_path)
+            if (!AppendNameRef(t_name))
+                return false;
+        return true;
+    }
+    
+private:
+    bool AppendNameRef(MCNameRef p_name)
+    {
+        if (m_container.m_path_length < sizeof(m_container.m_short_path) / sizeof(m_container.m_short_path[0]))
+        {
+            m_container.m_short_path[m_container.m_path_length++] = MCValueRetain(p_name);
+        }
+        else if (m_container.m_path_length == sizeof(m_container.m_short_path) / sizeof(m_container.m_short_path[0]))
+        {
+            MCNameRef *t_long_path;
+            if (!MCMemoryNewArray(sizeof(m_container.m_short_path) / sizeof(m_container.m_short_path[0]) + 1, t_long_path, m_container.m_path_length))
+                return false;
+            MCMemoryCopy(t_long_path, m_container.m_short_path, sizeof(m_container.m_short_path));
+            t_long_path[m_container.m_path_length - 1] = MCValueRetain(p_name);
+            m_container.m_long_path = t_long_path;
+        }
+        else
+        {
+            if (!MCMemoryResizeArray(m_container.m_path_length + 1, m_container.m_long_path, m_container.m_path_length))
+                return false;
+            m_container.m_long_path[m_container.m_path_length - 1] = MCValueRetain(p_name);
+        }
+        return true;
+    }
+    
+    MCContainer& m_container;
+};
+
 // Resolve references to the appropriate element refered to by this Varref.
 bool MCVarref::resolve(MCExecContext& ctxt, MCContainer& r_container)
 {
+    MCAssert(r_container.m_variable == nullptr && r_container.m_path_length == 0);
+    
     if (dimensions == 0 && !isparam)
-        return MCContainer::createwithvariable(fetchvar(ctxt), r_container);
+    {
+        r_container.m_variable = fetchvar(ctxt);
+        return true;
+    }
 
 	MCExpression **t_dimensions;
 	if (dimensions == 1)
@@ -1379,73 +1429,31 @@ bool MCVarref::resolve(MCExecContext& ctxt, MCContainer& r_container)
 	else
 		t_dimensions = exps;
     
+    MCContainerBuilder t_builder(r_container);
+    
     // AL-2014-08-20: [[ ArrayElementRefParams ]] If the Varref refers to a container then
     //  resolving the path requires appending the new dimensions to the old path
     MCSpan<MCNameRef> t_old_path = getpath(ctxt);
-    MCAutoNameRefArray t_path;
-    if (!t_path . New(t_old_path.size()))
-        return false;
-
-    for (uindex_t i = 0; i < t_old_path.size(); i++)
-        t_path[i] = MCValueRetain(t_old_path[i]);
-    
-    auto t_push_value_as_name = [&](MCValueRef p_value) {
-        MCNewAutoNameRef t_key;
-        if (!ctxt . ConvertToName(p_value, &t_key))
-        {
-            ctxt . LegacyThrow(EE_VARIABLE_BADINDEX);
+    if (t_old_path.length() > 0)
+        if (!t_builder.AppendNameSpan(t_old_path))
             return false;
-        }
-        
-        if (!t_path.Push(*t_key))
-            return false;
-        
-        MCValueRetain(*t_key);
-        return true;
-    };
-    
+            
     for(uindex_t i = 0; i < dimensions; i++)
 	{
         MCAutoValueRef t_value;
         if (!ctxt . EvalExprAsValueRef(t_dimensions[i], EE_VARIABLE_BADINDEX, &t_value))
             return false;
         
-        MCAutoArrayRef t_array;
-        if (ctxt . ConvertToArray(*t_value, &t_array)
-                && !MCArrayIsEmpty(*t_array))
-        {
-            if (!MCArrayIsSequence(*t_array))
-            {
-                ctxt . LegacyThrow(EE_VARIABLE_BADINDEX);
-                return false;
-            }
+        if (!t_builder.AppendValueRef(ctxt, *t_value))
+            return false;
+    }
 
-            uindex_t t_length = MCArrayGetCount(*t_array);
-            for(uindex_t t_index = 1; t_index <= t_length; t_index += 1)
-            {
-                MCValueRef t_value_fetched;
-                if (!MCArrayFetchValueAtIndex(*t_array, t_index, t_value_fetched))
-                    return false;
-
-                if (!t_push_value_as_name(t_value_fetched))
-                    return false;
-            }
-        }
-        else
-        {
-            if (!t_push_value_as_name(*t_value))
-                return false;
-        }
-	}
-
-    MCNameRef *t_path_values = nullptr;
-    uindex_t t_count = 0;
-    t_path . Take(t_path_values, t_count);
-    
-    if (isparam)
-        return MCContainer::copywithpath(fetchcontainer(ctxt), t_path_values, t_count, r_container);
+    if (!isparam)
+        r_container.m_variable = fetchvar(ctxt);
     else
-        return MCContainer::createwithpath(fetchvar(ctxt), t_path_values, t_count, r_container);
+        r_container.m_variable = fetchcontainer(ctxt)->getvar();
+    
+    return true;
 }
 
 MCSpan<MCNameRef> MCVarref::getpath(MCExecContext& ctxt)