… benchmarks/_benchmarkrunner.livecodescript
… benchmarks/lcs/control/array.livecodescript
… benchmarks/lcs/control/the_adventures_of_sherlock_holmes.txt
… libfoundation/src/foundation-array.cpp
… libfoundation/src/foundation-name.cpp
… libfoundation/src/foundation-number.cpp
… libfoundation/src/foundation-private.h
… libfoundation/test/test_memory.cpp