This patch replaces an always dynamic 'path' array in MCContainer with a small static one (currently max 4 elements) which switches to a dynamic one when a longer path is needed.

This reduces the number of memory allocations, and extensions needed to resolve array elements where there are less than the short path length elements.

Surprisingly this appears to cause at least a 1.5-2x speed up for most benchmarks in that file.
```
develop-9.0 (HEAD):

Running ArrayNameKeys...
	Store[] 187 ms
	Fetch[] 125 ms
	Store[][] 286 ms
	Fetch[][] 197 ms
Running ArrayIndexKeys...
	Store[] 481 ms
	Fetch[] 432 ms
	Store[][] 894 ms
	Fetch[][] 766 ms
Running ArrayStringKeys...
	Store[] 236 ms
	Fetch[] 154 ms
	Store[][] 333 ms
	Fetch[][] 235 ms
Running VariableFetchLocal...
	Get Base 24 ms
	Get Base[<name>] 85 ms
	Get Base[<number>] 392 ms
	Get Base[<string>] 117 ms
	Get Base[<seq-1>] 184 ms
	Get Base[<name>][<name>] 124 ms
	Get Base[<number>][<number>] 721 ms
	Get Base[<string>][<string>] 177 ms
	Get Base[<seq-2>] 298 ms
Running VariableStoreLocal...
	Replace Into Base 27 ms
	Replace Into Base[<name>] 22 ms
	Replace Into Base[<number>] 99 ms
	Replace Into Base[<string>] 28 ms
	Replace Into Base[<seq-1>] 49 ms
	Replace Into Base[<name>][<name>] 34 ms
	Replace Into Base[<number>][<number>] 180 ms
	Replace Into Base[<string>][<string>] 45 ms
	Replace Into Base[<seq-2>] 76 ms
```
Compared to:
```
THIS PR:
Running ArrayNameKeys...
	Store[] 92 ms
	Fetch[] 43 ms
	Store[][] 155 ms
	Fetch[][] 63 ms
Running ArrayIndexKeys...
	Store[] 263 ms
	Fetch[] 220 ms
	Store[][] 526 ms
	Fetch[][] 435 ms
Running ArrayStringKeys...
	Store[] 111 ms
	Fetch[] 66 ms
	Store[][] 201 ms
	Fetch[][] 118 ms
Running VariableFetchLocal...
	Get Base 22 ms
	Get Base[<name>] 45 ms
	Get Base[<number>] 205 ms
	Get Base[<string>] 64 ms
	Get Base[<seq-1>] 144 ms
	Get Base[<name>][<name>] 72 ms
	Get Base[<number>][<number>] 405 ms
	Get Base[<string>][<string>] 112 ms
	Get Base[<seq-2>] 247 ms
Running VariableStoreLocal...
	Replace Into Base 26 ms
	Replace Into Base[<name>] 12 ms
	Replace Into Base[<number>] 51 ms
	Replace Into Base[<string>] 17 ms
	Replace Into Base[<seq-1>] 35 ms
	Replace Into Base[<name>][<name>] 16 ms
	Replace Into Base[<number>][<number>] 97 ms
	Replace Into Base[<string>][<string>] 29 ms
	Replace Into Base[<seq-2>] 60 ms
```

The only thing to figure out is how long short path should be.

Previously MCContainer was 32 bytes on 64-bit systems, with short path length as 4, this climbs to 48 bytes. 

Whilst most MCContainer instances are on-stack in leaf functions (the exec commands which manipulate variables) - they also form the argument list passed to handlers. Therefore, increasing the size of the structure will have an impact on memory usage for deeply nested call stacks.

I'd perhaps suggest that 64 bytes would be the largest size that we would want to consider - this would mean array paths of up to 6 in length wouldn't require an auxillary allocation to access.

Note: As a side-effect, it should also give meaning to sequence arrays in sequences arrays. i.e. `[ [ a, b ], [c, [ e, f ] ] ]` is treated the as [ a, b, c, d, e, f ] (although this needs to be tested).