@runrevmark 2018-08-28
This needs split up / cleaned up - but it now contains a couple of other optimizations.

The size of all core MCValue structs has been minimised (all the main types are now <= 32 bytes on 64-bit).

Additionally the notion of 'index name' has been introduced more thoroughly (taken from previous attempts at this) - even the relatively simple implementation (which essentially replaces sprintf with a custom integer string generation function) appears to have made somewhat significant performance gains:

```
Running ArrayNameKeys...
	Store[] 90 ms
	Fetch[] 41 ms
	Store[][] 150 ms
	Fetch[][] 59 ms
Running ArrayIndexKeys...
	Store[] 147 ms
	Fetch[] 89 ms
	Store[][] 248 ms
	Fetch[][] 160 ms
Running ArrayStringKeys...
	Store[] 118 ms
	Fetch[] 61 ms
	Store[][] 202 ms
	Fetch[][] 113 ms
Running VariableFetchLocal...
	Get Base 22 ms
	Get Base[<name>] 45 ms
	Get Base[<number>] 88 ms
	Get Base[<string>] 59 ms
	Get Base[<seq-1>] 75 ms
	Get Base[<name>][<name>] 69 ms
	Get Base[<number>][<number>] 157 ms
	Get Base[<string>][<string>] 101 ms
	Get Base[<seq-2>] 103 ms
Running VariableStoreLocal...
	Replace Into Base 27 ms
	Replace Into Base[<name>] 11 ms
	Replace Into Base[<number>] 23 ms
	Replace Into Base[<string>] 16 ms
	Replace Into Base[<seq-1>] 18 ms
	Replace Into Base[<name>][<name>] 16 ms
	Replace Into Base[<number>][<number>] 42 ms
	Replace Into Base[<string>][<string>] 26 ms
	Replace Into Base[<seq-2>] 25 ms
```

Of course - this might not actually be working correctly (as I've not added any tests yet!) - but the benchmarks at least appear to work...